#1010 [REFACTOR] report 관련 constant, lib들을 report 폴더로 이동
