+++ src/feature/report/lib/parseReportType.js
